requirements.txt
# Core dependencies
pandas>=2.0.0
python-dotenv>=1.0.0

# Video/Audio processing
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0

# OpenAI APIs
openai>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
aiohttp>=3.9.0
asyncio>=3.4.3
